[batch] unify clients
