+++ batch/batch/__init__.py
@@ -1,10 +1,9 @@
-from . import api, client, aioclient
+from . import api, client
 from .poll_until import poll_until
 
 
 __all__ = [
     'client',
-    'aioclient',
     'api',
     'poll_until'
 ]
