--- batch/test/diamond_dag.yml
@@ -1,27 +1,0 @@
-- type: execute
-  name: head
-  spec:
-    image: alpine:3.8
-    command: ["echo", "head"]
-- type: execute
-  name: left
-  spec:
-    image: alpine:3.8
-    command: ["echo", "left"]
-  dependsOn:
-    - head
-- type: execute
-  name: right
-  spec:
-    image: alpine:3.8
-    command: ["echo", "right"]
-  dependsOn:
-    - head
-- type: execute
-  name: tail
-  spec:
-    image: alpine:3.8
-    command: ["echo", "tail"]
-  dependsOn:
-    - left
-    - right
